check uploader before deleting chapter

+++ backend/controllers/chapter.py
@@ -3,12 +3,11 @@ import time
 from app import db
 from cloudinary import api, uploader
 from flask import Blueprint, request
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import get_jwt_identity, jwt_required
 from models.chapter import Chapter
 from models.manga import Manga
 
 chapter_bp = Blueprint('chapter_bp', __name__)
-
 
 @chapter_bp.route('/chapter', methods=['POST'])
 @jwt_required()
@@ -49,14 +48,23 @@ def chapter_pages(chapter_id):
     pages = api.resources(type="upload", prefix=chapter.pages, max_results=100)
     return pages, 200
 
+
 @chapter_bp.route('/chapter/<string:chapter_id>', methods=['DELETE'])
 @jwt_required()
 def delete_chapter(chapter_id):
     chapter = Chapter.query.filter_by(id=chapter_id).first()
     if (not chapter):
         return {'message': 'Chapter not found'}, 400
+        
+    # Get manga to check if it is the uploader
+    manga = Manga.query.filter_by(id=chapter.manga_id).first()
 
+    if(manga.uploaded_by != get_jwt_identity()):
+        return {'message': 'Not Uploader'}, 403
+
+    # Delete images inside folder
     api.delete_resources_by_prefix(chapter.pages)
+    # Delete folder (has to be empty first)
     api.delete_folder(chapter.pages)
 
     db.session.delete(chapter)
